Break event warning icons out into separate component
