+++ src/features/campaigns/components/ActivityList/items/EventListItem.tsx
@@ -1,16 +1,12 @@
 import { FC } from 'react';
-import { Box, CircularProgress, Tooltip } from '@mui/material';
 import {
-  EmojiPeople,
   EventOutlined,
-  FaceRetouchingOff,
   Group,
-  MailOutline,
   PlaceOutlined,
   ScheduleOutlined,
 } from '@mui/icons-material';
 
-import { useMessages } from 'core/i18n';
+import EventWarningIcons from 'features/events/components/EventWarningIcons';
 import useModel from 'core/useModel';
 import ZUIIconLabelRow from 'zui/ZUIIconLabelRow';
 import ZUITimeSpan from 'zui/ZUITimeSpan';
@@ -19,19 +15,15 @@ import EventDataModel, {
   EventState,
 } from 'features/events/models/EventDataModel';
 
-import messageIds from 'features/campaigns/l10n/messageIds';
-
 interface EventListeItemProps {
   orgId: number;
   eventId: number;
 }
 
 const EventListItem: FC<EventListeItemProps> = ({ eventId, orgId }) => {
-  const messages = useMessages(messageIds);
   const model = useModel((env) => new EventDataModel(env, orgId, eventId));
   const state = model.state;
   const data = model.getData().data;
-  const participants = model.getParticipants();
 
   if (!data) {
     return null;
@@ -46,51 +38,6 @@ const EventListItem: FC<EventListeItemProps> = ({ eventId, orgId }) => {
     color = STATUS_COLORS.BLUE;
   }
 
-  const icons: JSX.Element[] = [
-    <AlertSlot key="contact" />,
-    <AlertSlot key="signups" />,
-    <AlertSlot key="reminders" />,
-  ];
-
-  if (participants.data) {
-    if (!data.contact) {
-      icons[0] = (
-        <AlertSlot
-          key="contact"
-          icon={<FaceRetouchingOff color="error" />}
-          tooltip={messages.activityList.eventItem.contact()}
-        />
-      );
-    }
-
-    const numBooked = participants.data.length;
-    if (data.num_participants_available > numBooked) {
-      icons[1] = (
-        <AlertSlot
-          key="signups"
-          icon={<EmojiPeople color="error" />}
-          tooltip={messages.activityList.eventItem.signups()}
-        />
-      );
-    }
-
-    const reminded = participants.data.filter((p) => !!p.reminder_sent);
-    const numReminded = reminded.length;
-    if (numReminded < participants.data.length) {
-      icons[2] = (
-        <AlertSlot
-          key="reminders"
-          icon={<MailOutline color="error" />}
-          tooltip={messages.activityList.eventItem.reminders({
-            numMissing: participants.data.length - numReminded,
-          })}
-        />
-      );
-    }
-  } else {
-    icons.push(<CircularProgress />);
-  }
-
   return (
     <ActivityListItem
       color={color}
@@ -103,7 +50,7 @@ const EventListItem: FC<EventListeItemProps> = ({ eventId, orgId }) => {
       href={`/organize/${orgId}/projects/${
         data.campaign?.id ?? 'standalone'
       }/events/${eventId}`}
-      meta={<Box display="flex">{icons}</Box>}
+      meta={<EventWarningIcons model={model} />}
       PrimaryIcon={EventOutlined}
       SecondaryIcon={Group}
       subtitle={
@@ -129,23 +76,6 @@ const EventListItem: FC<EventListeItemProps> = ({ eventId, orgId }) => {
       }
       title={data.title || data.activity.title}
     />
-  );
-};
-
-const AlertSlot: FC<{
-  icon?: JSX.Element;
-  tooltip?: string;
-}> = ({ icon, tooltip }) => {
-  return (
-    <Box width="1.6em">
-      {tooltip ? (
-        <Tooltip arrow title={tooltip}>
-          <Box>{icon}</Box>
-        </Tooltip>
-      ) : (
-        icon
-      )}
-    </Box>
   );
 };
